rename to ChartHook, chart options, monitor components
